requirements.txt
tensorflow==2.4.0
streamlit==0.73.1
numpy==1.19.4
pillow==8.0.1